post and delete

+++ controller/structureContentController.js
@@ -2,6 +2,11 @@ import asyncHandler from "express-async-handler"
 import mongoose from "mongoose"
 import StructureContent from "../model/structureContentModel.js";
 import moment from "moment";
+import { GridFSBucket, MongoClient, ObjectId } from "mongodb";
+
+const url = 'mongodb://127.0.0.1:27017/'
+const baseUrl = ' http://localhost:4242/api/image/upload/'
+
 
 // @desc    Fetch Contents
 // route    GET /api
@@ -81,14 +86,21 @@ const sortContent = asyncHandler(async (req, res) => {
 //@access   Public
 const createContent = asyncHandler(async (req, res) => {
     const { title, content } = req.body
+    const { file } = req
 
     try {
+        if (!file) {
+            throw new Error('Field name "image" missing in form data');
+        }
+
         // Clear the database first
         await StructureContent.deleteMany();
 
         var createContent = new StructureContent({
             title,
-            content
+            content,
+            id: file.id,
+            images: file.filename
         })
 
         var singleContent = await createContent.save()
@@ -130,8 +142,13 @@ const updateContent = asyncHandler(async (req, res) => {
 // route    DELETE /api/content/:id
 //@access   Public
 const deleteContent = asyncHandler(async (req, res) => {
+    const { id, Id } = req.body
 
-    const { id } = req.body
+    const mongoClient = new MongoClient(url);
+    await mongoClient.connect();
+
+    const database = mongoClient.db('test'); // Adjust database name if needed
+    const bucket = new GridFSBucket(database, { bucketName: 'images' });
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(404).json({ error: "No such document" })
@@ -143,6 +160,8 @@ const deleteContent = asyncHandler(async (req, res) => {
         if (!content) {
             return res.status(404).json({ error: "Content not found" });
         }
+
+        await bucket.delete(new ObjectId(Id));
 
         const result = await StructureContent.findByIdAndDelete({ _id: id })
